Update tres_claw_switch.c
The file was changed to use the SGProps format.  It no longer needs pointers to global defs.  The SGProp and local variables are created before the functions start, and they are initially set in the init_mission function.  At the end of the program after the local variables have been adjusted, these local variables are used to set the variable values in SGProps.