Update and rename EKF_15state_quat.c to EKF_15state_quat.cxx
This was changed into the SGProps format and it no longer relies on globaldefs.h.   The SGProperty variables and local variables are created at the very beginning before the functions.  Then they are initialized within init_nav.  At the end of init_nav and get_nav, the output variables for SGProps are set using their corresponding local variables.  The file was changed to a c++ file to work with props.hxx.