Create extendData.h5
This is a working code that creates and extends a dataset.  This code was modified from the original h5_extend_write.c example on the HDF5 website.  One of the main differences in this code from the original example is the loop makes it possible to append the file for any set amount of times.  This is just a file to start with.  This code still needs to be edited so it uses inputed data from the aircraft (rather than setting data array values at the beginning) and so it will stop running the code when the aircraft stops collecting data.  In the end, this code should extend it's memory size during the flight to take in more data value and then should stop at the end of the flight.